% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculate Performance Metrics}
\usage{
calc_metrics(gains = NULL, metrics = c("mean", "sd", "growth", "cagr",
  "mdd", "sharpe", "sortino", "alpha", "alpha.annualized", "beta",
  "r.squared", "pearson", "spearman", "auto.pearson", "auto.spearman"),
  prices = NULL, tickers = NULL, ..., benchmark = "SPY")
}
\arguments{
\item{gains}{Data frame with one column of gains for each investment and a 
date variable named Date.}

\item{metrics}{Character vector specifying metrics to calculate.}

\item{prices}{Data frame with one column of prices for each investment and a 
date variable named Date.}

\item{tickers}{Character vector of ticker symbols that Yahoo! Finance 
recognizes, if you want to download data on the fly.}

\item{...}{Arguments to pass along with \code{tickers} to 
\code{\link{load_gains}}.}

\item{benchmark}{Character string specifying which fund to use as a 
benchmark for metrics that require one.}
}
\value{
Data frame with performance metrics for each investment.
}
\description{
Useful for comparing funds on one or more metrics.
}
\examples{
\dontrun{
# Calculate performance metrics for FANG stocks
calc_metrics(tickers = c("FB", "AAPL", "NFLX", "GOOG"))
}


}
