% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics_123.R
\name{calc_metrics_123}
\alias{calc_metrics_123}
\title{Calculate Performance Metrics for Any Combination of Individual Funds,
2-Fund Portfolios, and 3-Fund Portfolios}
\usage{
calc_metrics_123(gains = NULL, metrics = c("mean", "sd"),
  tickers = NULL, ..., step.along = 2.5, step.between = 2.5,
  prices = NULL, benchmark = "SPY")
}
\arguments{
\item{gains}{Data frame with a date variable named Date and one column of
gains for each investment.}

\item{metrics}{Character vector specifying metrics to calculate. See
\code{?calc_metrics} for choices.}

\item{tickers}{Character vector of ticker symbols, where the first three are
are a 3-fund set, the next three are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{step.along}{Numeric value controlling allocation increments along each
curve. Only applies to 2- and 3-fund sets.}

\item{step.between}{Numeric value controlling allocation increments between
each curve. Only applies to 3-fund sets.}

\item{prices}{Data frame with a date variable named Date and one column of
prices for each investment.}

\item{benchmark}{Character string specifying which fund to use as a
benchmark for metrics that require one.}
}
\value{
Data frame with performance metrics for each portfolio at each allocation.
}
\description{
Integrates \code{calc_metrics}, \code{calc_metrics_2funds}, and
\code{calc_metrics_3funds} into a single function, so you can compare
strategies of varying complexities.
}
\examples{
\dontrun{
# Calculate CAGR vs. max drawdown for BRK-B, SPY/TLT, and VWEHX/VBLTX/VFINX
df <- calc_metrics_123(
  tickers = list("BRK-B", c("SPY", "TLT"), c("VWEHX", "VBLTX", "VFINX")),
  metrics = c("cagr", "mdd")
)
head(df)

# To plot, just pipe into plot_metrics_123
df \%>\%
  plot_metrics_123()

# Or bypass calc_metrics_123 altogether
plot_metrics_123(
  formula = cagr ~ mdd,
  tickers = list("BRK-B", c("SPY", "TLT"), c("VWEHX", "VBLTX", "VFINX"))
)
}


}
