% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics_2funds.R
\name{calc_metrics_2funds}
\alias{calc_metrics_2funds}
\title{Calculate Performance Metrics for Two-Fund Portfolios with Varying 
Allocations}
\usage{
calc_metrics_2funds(gains = NULL, metrics = c("mean", "sd"),
  tickers = NULL, ..., prices = NULL, benchmark = "SPY",
  ref.tickers = "SPY")
}
\arguments{
\item{gains}{Data frame with a date variable named Date and one column of 
gains for each investment.}

\item{metrics}{Character vector specifying metrics to calculate. See 
\code{?calc_metrics} for choices.}

\item{tickers}{Character vector of ticker symbols, where the first two are 
are a two-fund pair, the next two are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to 
\code{\link{load_gains}}.}

\item{prices}{Data frame with a date variable named Date and one column of 
prices for each investment.}

\item{benchmark}{Character string specifying which fund to use as a 
benchmark for metrics that require one.}

\item{ref.tickers}{Character vector of ticker symbols to include.}
}
\value{
Depending on \code{return}, a \code{\link[ggplot2]{ggplot}} object, a data
frame, or a list containing both.
}
\description{
Useful for assessing the characteristics of two-fund portfolios.
}
\examples{
\dontrun{
# Plot mean vs. SD for UPRO/VBLTX, and compare to SPY
plot_metrics_2funds(mean ~ sd, c("UPRO", "VBLTX"))

# Plot CAGR vs. MDD for AAPL/GOOG and FB/TWTR
plot_metrics_2funds(cagr ~ mdd, tickers = c("AAPL", "GOOG", "FB", "TWTR"))

# Plot Sharpe ratio vs. allocation for SPY/TLT
plot_metrics_2funds(sharpe ~ allocation, tickers = c("SPY", "TLT"))
}

}
