% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics_3funds.R
\name{calc_metrics_3funds}
\alias{calc_metrics_3funds}
\title{Calculate Performance Metrics for 3-Fund Portfolios with Varying Allocations}
\usage{
calc_metrics_3funds(gains = NULL, metrics = c("mean", "sd"),
  tickers = NULL, ..., step1 = 2.5, step2 = step1, prices = NULL,
  benchmark = "SPY", ref.tickers = NULL)
}
\arguments{
\item{gains}{Data frame with a date variable named Date and one column of
gains for each investment.}

\item{metrics}{Character vector specifying metrics to calculate. See
\code{?calc_metrics} for choices.}

\item{tickers}{Character vector of ticker symbols, where the first three are
are a 3-fund set, the next three are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{step1}{Numeric value controlling allocation increments for first fund
of each set.}

\item{step2}{Numeric value controlling allocation increments for the second
and third fund in each set.}

\item{prices}{Data frame with a date variable named Date and one column of
prices for each investment.}

\item{benchmark}{Character string specifying which fund to use as a
benchmark for metrics that require one.}

\item{ref.tickers}{Character vector of ticker symbols to include.}
}
\value{
Data frame with performance metrics for each portfolio at each allocation.
}
\description{
Useful for assessing the characteristics of 3-fund portfolios.
}
\examples{
\dontrun{
# Calculate CAGR and max drawdown for UPRO/VBLTX/VWEHX
df <- calc_metrics_3funds(metrics = c("cagr", "mdd"), tickers = c("UPRO", "VBLTX", "VWEHX"))
head(df)

# To plot, just pipe into plot_metrics_3funds
df \%>\%
  plot_metrics_3funds()

# Or bypass calc_metrics_3funds altogether
plot_metrics_3funds(formula = cagr ~ mdd, tickers = c("UPRO", "VBLTX", "VWEHX"))
}


}
