% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Financial_Ratios.R
\name{filter_company_data}
\alias{filter_company_data}
\title{Quick Look at a Company's Financials}
\usage{
filter_company_data(tik)
}
\arguments{
\item{company_name}{Write in company name in quotes}
}
\value{
The stock ticker for the company if it is a public stock on NASDAQ or NYSE
}
\description{
This function will allow a use to look at several ratios of the company
}
\examples{
#Example usage:
company_data <- filter_company_data(tik = "AAPL")
company_data
}
