% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gains_rate.R
\name{gains_rate}
\alias{gains_rate}
\title{Calculate Growth Rate from Sequence of Gains}
\usage{
gains_rate(gains, units.out = NULL)
}
\arguments{
\item{gains}{Data frame with one column of gains for each investment (can be 
a numeric vector if there is only one).}

\item{units.out}{Numeric value specifying the number of units for growth 
rate calculation, if you want something other than total growth. For 
annualized growth rate, set to 252 if \code{gains} has daily gains, 12 if 
\code{gains} has monthly gains, etc.}
}
\value{
Numeric vector.
}
\description{
The formula is simply: \code{prod(gains + 1) - 1}. If \code{units.out} is 
specified, then it converts to x-unit growth rate.
}
\examples{
# Create vector of daily gains for a hypothetical stock
daily.gains <- c(-0.02, -0.01, 0.01, 0.02, 0.01)

# Overall growth is 0.95\%
gains_rate(daily.gains)

# Average daily growth is 0.19\%
gains_rate(daily.gains, 1)

# Corresponds to 61.0\% annual growth
gains_rate(daily.gains, 252)


}
