% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_stocks.R
\name{get_last_valid_trading_day}
\alias{get_last_valid_trading_day}
\title{Plot Multiple Stocks on One Plot}
\usage{
get_last_valid_trading_day(to_date)
}
\arguments{
\item{to_date}{The ending date of the plot, has to be done as MM/DD/YYYY}

\item{tickers}{The stock tickers that will be plotted}

\item{from_date}{The starting date of the plot, has to be done as MM/DD/YYYY}
}
\value{
One plot will be generated with stocks prices of selected stocks over a certain amount of time specified. One table will be generated giving the percent increase or descrease of the stock over the specified amount of time.
}
\description{
Allows user to see multiple stocks on one plot
}
\examples{
plot_multiple_stocks(c("PAR", "MSFT", "BA", "AAPL"), from_date = "01/05/2023", to_date = "01/01/2024")
}
