% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{highyield_etfs}
\alias{highyield_etfs}
\title{High-Yield ETFs from ETFdb.com}
\source{
\url{https://etfdb.com/etfdb-category/high-yield-bonds/}
}
\description{
High-Yield ETFs from ETFdb.com
}
