% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_metric.R
\name{label_metric}
\alias{label_metric}
\title{Convert Label back to Performance Metric}
\usage{
label_metric(label)
}
\arguments{
\item{label}{Character string.}
}
\value{
Character string.
}
\description{
For internal use only.
}
