\name{largest.etfs}
\alias{largest.etfs}
\docType{data}
\title{
List of 100 Largest Market Cap ETFs as of April 9, 2017
}
\description{
Simply a character vector of ticker symbols for ETFs.
}
\format{
  A character vector with 100 tickers.
}
\source{
\url{http://etfdb.com/compare/market-cap/}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}