\name{largest_etfs_4_9_17}
\alias{largest_etfs_4_9_17}
\docType{data}
\title{
Largest 100 Market Cap ETFs and Inception Dates as of April 9, 2017
}
\description{
Data frame where first column is ticker symbol and second column is inception
date.
}
\format{
  A 100-row data frame with 2 variables: \code{ticker} (character string) and
  \code{start.date} (date).
}
\source{
\url{http://etfdb.com/compare/market-cap/}
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}
