% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prices.R
\name{load_prices}
\alias{load_prices}
\title{Download Historical Prices}
\usage{
load_prices(tickers, intercepts = NULL, slopes = NULL,
  from = "1950-01-01", to = Sys.Date(), time.scale = "daily",
  preto.days = NULL, prefrom.days = NULL, initial = NULL,
  mutual.start = FALSE, mutual.end = TRUE, anchor = FALSE,
  drop.anyNA = FALSE)
}
\arguments{
\item{tickers}{Character vector of ticker symbols that Yahoo! Finance 
recognizes.}

\item{intercepts}{Numeric vector of values to add to daily gains for each 
fund.}

\item{slopes}{Numeric vector of values to multiply daily gains for each fund 
by. Slopes are multiplied prior to adding intercepts.}

\item{from}{Date or character string, e.g. \code{"2015-01-15"}.}

\item{to}{Date or character string, e.g. \code{"2018-12-31"}.}

\item{time.scale}{Character string. Choices are \code{"daily"}, 
\code{"monthly"}, and \code{"yearly"}.}

\item{preto.days}{Numeric value. If specified, function returns prices for 
\code{preto.days} trading days prior to \code{to}. For example, to load the 
most recent 50 closing prices, leave \code{to} and \code{time.scale} as the 
defaults and set \code{preto.days = 50}.}

\item{prefrom.days}{Numeric value. If specified, function returns prices for 
\code{prefrom.days} trading days prior to \code{from}. Useful when you want 
to test a trading strategy starting on a particular date, but the strategy 
requires data leading up to that date (e.g. trailing beta).}

\item{initial}{Numeric value specifying what value to scale initial prices 
to.}

\item{mutual.start}{Logical value for whether to start on the first day of
the funds' mutual lifetimes.}

\item{mutual.end}{Logical value for whether to end on the last day of the 
funds' mutual lifetimes.}

\item{anchor}{Logical value for whether to anchor the starting price for each 
fund to the price of the longest-running fund on that day. Useful for 
visualizing funds' entire histories while also fairly comparing them over 
their mutual lifetimes. Only used if \code{mutual.start = FALSE}.}

\item{drop.anyNA}{Logical value for whether to drop dates on which prices are
missing for any of the funds.}
}
\value{
Data frame with closing prices for each fund.
}
\description{
Downloads historical prices for specified tickers from Yahoo! Finance, with 
various options. Relies heavily on the \pkg{quantmod} package.
}
\examples{
\dontrun{
# Load prices for Netflix and Amazon over their mutual lifetimes
prices <- load_prices(c("NFLX", "AMZN"))
}


}
\references{
Jeffrey A. Ryan and Joshua M. Ulrich (2019). quantmod: Quantitative Financial 
Modelling Framework. R package version 0.4-15. 
\url{https://CRAN.R-project.org/package=quantmod}
}
