% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_decimals.R
\name{metric_decimals}
\alias{metric_decimals}
\title{Get Number of Decimals for Performance Metric}
\usage{
metric_decimals(metric)
}
\arguments{
\item{metric}{Character string.}
}
\value{
Character string.
}
\description{
Mainly a helper function.
}
