% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_label.R
\name{metric_label}
\alias{metric_label}
\title{Get Label for Performance Metric}
\usage{
metric_label(metric)
}
\arguments{
\item{metric}{Character string.}
}
\value{
Character string.
}
\description{
Mainly a helper function.
}
