% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_title.R
\name{metric_title}
\alias{metric_title}
\title{Get Title for Performance Metric}
\usage{
metric_title(metric)
}
\arguments{
\item{metric}{Character string.}
}
\value{
Character string.
}
\description{
Mainly a helper function.
}
