% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_units.R
\name{metric_units}
\alias{metric_units}
\title{Get Units for Performance Metric}
\usage{
metric_units(metric)
}
\arguments{
\item{metric}{Character string.}
}
\value{
Character string.
}
\description{
Mainly a helper function.
}
