% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Calculate Various Performance Metrics}
\usage{
metrics(tickers = NULL, ..., gains = NULL, prices = NULL,
  perf.metrics = c("mean", "sd", "growth", "cagr", "mdd", "sharpe", "sortino",
  "alpha", "beta", "r.squared", "pearson", "spearman", "auto.pearson",
  "auto.spearman"))
}
\arguments{
\item{tickers}{Character vector of ticker symbols, if you wish to load data 
on the fly rather than specify \code{gains} or \code{prices}.}

\item{...}{Arguments to pass along with \code{tickers} to 
\code{\link{load_gains}}.}

\item{gains}{Numeric matrix of gains, with 1 column for each investment.}

\item{prices}{Numeric matrix of prices, with 1 column for each investment.}

\item{perf.metrics}{Character vector specying metrics to calculate. If at 
least 1 metric requires a benchmark, the first investment is used.}
}
\value{
List containing a data frame with the performance metrics and a correlation 
matrix for gains for the various investments.
}
\description{
Calculates variety of performance metrics from prices or gains.
}
\examples{
# Calculate performance metrics for leveraged ETFs SSO and UPRO, using SPY 
# as benchmark for alpha and beta - "on the fly" method
metrics1 <- metrics(tickers = c("SPY", "SSO", "UPRO"))

# Calculate same performance metrics, but specify gains input
gains <- load_gains(tickers = c("SPY", "SSO", "UPRO"))
metrics2 <- metrics(gains = gains)

# Results are identical
all.equal(metrics1$perf.metrics, metrics2$perf.metrics)


}
\references{
Ryan, J.A. and Ulrich, J.M. (2017) quantmod: Quantitative Financial Modelling 
Framework. R package version 0.4-12, 
\url{https://CRAN.R-project.org/package=quantmod}.
}
