% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gains.R
\name{plot_gains}
\alias{plot_gains}
\title{Plot Gains for One Investment vs. Another}
\usage{
plot_gains(
  formula = NULL,
  ...,
  gains = NULL,
  prices = NULL,
  poly_order = 1,
  plotly = FALSE,
  title = NULL,
  base_size = 16,
  return = "plot"
)
}
\arguments{
\item{formula}{Formula, e.g. \code{SSO + UPRO ~ SPY} to plot gains for SSO
and UPRO vs. SPY.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{gains}{Data frame with one column of gains for each investment
mentioned in \code{formula}. If unspecified, function downloads historical
gains internally.}

\item{prices}{Data frame with one column of prices for each investment
mentioned in \code{formula}.}

\item{poly_order}{Numeric value specifying the polynomial order for linear
regression, e.g. \code{1} for simple linear regression or \code{2} for
linear regression with first- and second-order terms.}

\item{plotly}{Logical value for whether to convert the
\code{\link[ggplot2]{ggplot}} to a \code{\link[plotly]{plotly}} object
internally. Note that legend displaying regression estimates will disappear
if you choose this option.}

\item{title}{Character string.}

\item{base_size}{Numeric value.}

\item{return}{Character string specifying what to return. Choices are
\code{"plot"}, \code{"data"}, and \code{"both"}.}
}
\value{
In addition to the graph, a list containing fitted linear regression models
returned by \code{\link[stats]{lm}} for each investment vs. the benchmark.
}
\description{
Useful for visualizing how two investments behave relate to each other, or
how several investments behave relative to the same benchmark.
}
\examples{
\dontrun{
# Plot daily gains for SSO and UPRO vs. VFINX
p <- plot_gains(SSO + UPRO ~ VFINX)
}

}
\references{
Jeffrey A. Ryan and Joshua M. Ulrich (2019). quantmod: Quantitative Financial
Modelling Framework. R package version 0.4-15.
\url{https://CRAN.R-project.org/package=quantmod}
}
