% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot Investment Growth}
\usage{
plot_growth(prices = NULL, tickers = NULL, ..., gains = NULL,
  initial = 10000, return = "plot")
}
\arguments{
\item{prices}{Data frame with one column of prices for each investment and a 
date variable named Date.}

\item{tickers}{Character vector of ticker symbols that Yahoo! Finance 
recognizes, if you want to download data on the fly.}

\item{...}{Arguments to pass along with \code{tickers} to 
\code{\link{load_gains}}.}

\item{gains}{Data frame with one column of gains for each investment and a 
date variable named Date.}

\item{initial}{Numeric value specifying value to scale initial prices to.}

\item{return}{Character string specifying what to return. Choices are 
\code{"plot"}, \code{"data"}, and \code{"both"}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Useful for comparing the performance of several investments, over their full 
histories or mutual lifetimes.
}
\examples{
\dontrun{
# Plot growth of $10k in VFINX and BRK-B
plot_growth(tickers = c("VFINX", "BRK-B"))
}

}
