% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot One Performance Metric (Sorted Bar Plot) or One vs. Another
(Scatterplot)}
\usage{
plot_metrics(metrics = NULL, formula = cagr ~ mdd, tickers = NULL,
  ..., gains = NULL, prices = NULL, benchmark = "SPY",
  y.benchmark = benchmark, x.benchmark = benchmark, ggplotly = FALSE,
  title = NULL, return = "plot")
}
\arguments{
\item{metrics}{"Long" data frame with Fund column and column for each metric
you want to plot. Typically the result of a prior call to
\code{\link{calc_metrics}}.}

\item{formula}{Formula specifying what to plot, e.g. \code{cagr ~ mdd} for
CAGR vs. MDD, \code{cagr ~ .} for just CAGR, or \code{. ~ mdd} for just MDD.
See \code{?calc_metrics} for list of metrics to choose from.}

\item{tickers}{Character vector of ticker symbols that Yahoo! Finance
recognizes, if you want to download data on the fly.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{gains}{Data frame with one column of gains for each investment and a
date variable named Date.}

\item{prices}{Data frame with one column of prices for each investment and a
date variable named Date.}

\item{benchmark}{Character string specifying which fund to use as a
benchmark for metrics that require one.}

\item{y.benchmark}{Character string specifying which fund to use as benchmark
for y-axis metric.}

\item{x.benchmark}{Character string specifying which fund to use as benchmark
for x-axis metric.}

\item{title}{Character string. Only really useful if you're going to set
\code{plotly = TRUE}, otherwise you can change the title, axes, etc.
afterwards.}

\item{return}{Character string specifying what to return. Choices are
\code{"plot"}, \code{"data"}, and \code{"both"}.}

\item{plotly}{Logical value for whether to convert the
\code{\link[ggplot2]{ggplot}} to a \code{\link[plotly]{plotly}} object
internally.}
}
\value{
Depending on \code{return}, a \code{\link[ggplot2]{ggplot}}, a data frame
with the source data, or a list containing both.
}
\description{
Useful for comparing funds on one or more metrics.
}
\examples{
\dontrun{
# Plot CAGR vs. MDD for SPY and BRK-B, downloading data on the fly
plot_metrics(formula = cagr ~ mdd, tickers = c("SPY", "BRK-B"))

# Plot Sharpe ratio for FANG stocks, in more explicit steps with piping
plot_metrics(formula = . ~ sharpe, tickers = c("FB", "AAPL", "NFLX", "GOOG"))
}


}
\references{
Jeffrey A. Ryan and Joshua M. Ulrich (2019). quantmod: Quantitative Financial
Modelling Framework. R package version 0.4-15.
\url{https://CRAN.R-project.org/package=quantmod}
}
