% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics_3funds.R
\name{plot_metrics_3funds}
\alias{plot_metrics_3funds}
\title{Plot One Performance Metric vs. Another for 3-Fund Portfolios}
\usage{
plot_metrics_3funds(metrics = NULL, formula = mean ~ sd,
  tickers = NULL, ..., step = 1, gains = NULL, prices = NULL,
  benchmark = "SPY", y.benchmark = benchmark,
  x.benchmark = benchmark, ref.tickers = NULL, plotly = FALSE,
  title = NULL, base_size = 16, label_size = 6, return = "plot")
}
\arguments{
\item{metrics}{Data frame with Fund column and column for each metric you
want to plot. Typically the result of a prior call to
\code{\link{calc_metrics_3funds}}.}

\item{formula}{Formula specifying what to plot, e.g. \code{mean ~ sd},
\code{cagr ~ mdd}, or \code{sharpe ~ allocation}. See \code{?calc_metrics}
for list of metrics to choose from (\code{"allocation"} is an extra option
here). If you specify \code{metrics}, default behavior is to use
\code{mean ~ sd} unless either is not available, in which case the first two
performance metrics that appear as columns in \code{metrics} are used.}

\item{tickers}{Character vector of ticker symbols, where the first three are
are a 3-fund set, the next three are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{step}{Numeric value specifying fund allocation increments.}

\item{gains}{Data frame with a date variable named Date and one column of
gains for each fund.}

\item{prices}{Data frame with a date variable named Date and one column of
prices for each fund.}

\item{benchmark, y.benchmark, x.benchmark}{Character string specifying which
fund to use as benchmark for metrics (if you request \code{alpha},
\code{alpha.annualized}, \code{beta}, or \code{r.squared}).}

\item{ref.tickers}{Character vector of ticker symbols to include on the
graph.}

\item{plotly}{Logical value for whether to convert the
\code{\link[ggplot2]{ggplot}} to a \code{\link[plotly]{plotly}} object
internally.}

\item{title}{Character string.}

\item{base_size}{Numeric value.}

\item{label_size}{Numeric value.}

\item{return}{Character string specifying what to return. Choices are
\code{"plot"}, \code{"data"}, and \code{"both"}.}
}
\value{
Depending on \code{return}, a \code{\link[ggplot2]{ggplot}} object, a data
frame, or a list containing both.
}
\description{
Useful for visualizing the behavior of one or several 3-fund portfolios,
e.g. by plotting a measure of growth vs. a measure of volatility.
}
\examples{
\dontrun{
# Plot mean vs. SD for UPRO/VBLTX/VWEHX
plot_metrics_3funds(
  formula = mean ~ sd,
  tickers = c("UPRO", "VBLTX", "VWEHX")
)

# Plot CAGR vs. max drawdown for FB/AAPL/NFLX and SPY/TLT/JNK
plot_metrics_3funds(
  formula = cagr ~ mdd,
  tickers = c("FB", "AAPL", "NFLX", "SPY", "TLT", "JNK")
)

# Plot Sharpe ratio vs. allocation for the same sets
plot_metrics_3funds(
  formula = sharpe ~ allocation,
  tickers = c("FB", "AAPL", "NFLX", "SPY", "TLT", "JNK")
)
}


}
