% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_stocks.R
\name{plot_multiple_stocks}
\alias{plot_multiple_stocks}
\title{Plot Multiple Stocks on One Plot}
\usage{
plot_multiple_stocks(tickers, from_date = NULL, to_date = NULL)
}
\arguments{
\item{tickers}{A vector of stock tickers to be plotted.}

\item{from_date}{The starting date of the plot, formatted as MM/DD/YYYY.}

\item{to_date}{The ending date of the plot, formatted as MM/DD/YYYY.}
}
\value{
A plot showing stock prices and a table with percentage changes over the specified time period.
}
\description{
Allows the user to see multiple stocks on one plot.
}
\examples{
plot_multiple_stocks(c("PAR", "MSFT", "BA", "AAPL"), from_date = "01/05/2023", to_date = "01/01/2024")
}
