% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_gains.R
\name{prices_gains}
\alias{prices_gains}
\title{Convert Sequence of Prices to Sequence of Gains}
\usage{
prices_gains(prices)
}
\arguments{
\item{prices}{Numeric vector of prices for one investment or data frame with 
one column for each investment and an optional Date variable.}
}
\value{
Numeric vector or data frame.
}
\description{
Converts sequence of prices to sequence of gains for one or more investments.
}
\examples{
\dontrun{
# Load 2017 prices for Netflix and Amazon, and calculate growth of $10k
prices <- load_prices(c("NFLX", "AMZN"), initial = 1000)

# Calculate gains
gains <- prices_gains(prices)
}


}
