% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_metric.R
\name{rolling_metric}
\alias{rolling_metric}
\title{Calculate Moving-Window Performance Metrics}
\usage{
rolling_metric(gains, metric = "mean", width = 50, units.year = 252,
  benchmark.gains = NULL)
}
\arguments{
\item{gains}{Numeric vector.}

\item{metric}{Character string.}

\item{width}{Integer value.}

\item{units.year}{Integer value.}

\item{benchmark.gains}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Mainly a helper function for \code{\link{plot_metrics_overtime}}.
}
