\name{sector.spdr.etfs}
\alias{sector.spdr.etfs}
\docType{data}
\title{
Sector SPDR ETFs and Inception Dates as of April 7, 2017
}
\description{
Data frame where first column is ticker symbol and second column is inception date.
}
\format{
  A 10-row data frame with 2 variables: \code{ticker} (character string) and \code{start.date} (date).
}
\source{
\url{http://www.sectorspdr.com/sectorspdr/sectors/performance}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}