% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{package}
\name{stocks}
\alias{stocks}
\alias{stocks-package}
\title{Stock Market Analysis}
\description{
Functions for analyzing stocks or other investments. Main features are
loading and aligning historical data for ticker symbols, calculating
performance metrics for individual funds or portfolios (e.g. compound
annualized growth rate, maximum drawdown, Sharpe/Sortino ratio), and creating
graphs. C++ code is used where possible to improve processing speed.
}
\details{
\tabular{ll}{
Package: \tab stocks \cr
Type: \tab Package \cr
Version: \tab 1.1.2 \cr
Date: \tab 2018-02-21 \cr
License: \tab GPL-3 \cr
}

See \href{https://cran.r-project.org/package=stocks}{CRAN documentation} for
full list of functions.
}
\references{
Ryan, J.A. and Ulrich, J.M. (2017) quantmod: Quantitative Financial Modelling 
Framework. R package version 0.4-12, 
\url{https://CRAN.R-project.org/package=quantmod}.

Acknowledgment: This material is based upon work supported by the
National Science Foundation Graduate Research Fellowship under Grant No.
DGE-0940903.
}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
