% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threefunds_graph.R
\name{threefunds_graph}
\alias{threefunds_graph}
\title{Graph One Performance Metric vs. Another for Three-Fund Portfolios as 
Allocation Varies}
\usage{
threefunds_graph(tickers = NULL, intercepts = NULL, slopes = NULL, ...,
  benchmark.tickers = NULL, reference.tickers = NULL,
  tickers.gains = NULL, benchmark.gains = NULL, reference.gains = NULL,
  step.data = 0.0025, step.points = 0.1, step.curves = 0.2,
  x.metric = "sd", y.metric = "mean", tickerlabel.offsets = NULL,
  reflabel.offsets = NULL, add.plot = FALSE, colors = NULL,
  plot.list = NULL, points.list = NULL, text.list = NULL,
  pdf.list = NULL, bmp.list = NULL, jpeg.list = NULL, png.list = NULL,
  tiff.list = NULL)
}
\arguments{
\item{tickers}{Character vector or matrix of ticker symbols, if you want to 
load data on the fly rather than specify \code{tickers.gains}. If a vector, 
the first through third elements are assumed to be first three-fund set, the 
fourth through sixth the second three-fund set, and so on. If a matrix, each 
three-row column should represent a different set.}

\item{intercepts}{Numeric vector or matrix of values to add to daily gains 
for each ticker. For example, if you have three tickers and want to simulate 
a 1\% annual expense ratio for only the third fund, you would set 
\code{intercepts = c(0, 0, convert.rate(-0.01, units.in = 252, units.out = 1))}. 
Like the \code{tickers} input, can be a vector divisible by three or a 
three-row matrix.}

\item{slopes}{Numeric vector or matrix of values to multiply daily gains for 
each ticker by. For example, if you have three tickers and want to simulate a 
2x leveraged version of the third fund, you would set 
\code{slopes = c(1, 1, 2)}. The slopes are multiplied prior to adding the 
intercepts. Like the \code{tickers} input, can be a vector divisible by three 
or a three-row matrix.}

\item{...}{Arguments to pass along with \code{tickers}, \code{intercepts}, 
and \code{slopes} to \code{\link{load_gains}} function, if you want to load 
data on the fly rather than specify \code{tickers.gains}.}

\item{benchmark.tickers}{Character vector of length 1 or 2 indicating ticker 
symbols for benchmark indexes. Only used if \code{x.metric} and/or 
\code{y.metric} require benchmark indexes to calculate.  For example, to plot 
correlation with SPY on the x-axis and correlation with TLT on the y-axis, 
set \code{x.metric = "pearson"}, \code{y.metric = "pearson2"} (i.e. Pearson 
correlation with 2nd benchmark), and 
\code{benchmark.tickers = c("SPY", "TLT")}.}

\item{reference.tickers}{Character vector of ticker symbols to include on
graph as data points for comparative purposes.}

\item{tickers.gains}{Numeric matrix of gains, where each column has gains for 
a particular fund. The first through third columns are assumed to be the 
first three-fund set, fourth through sixth the second, and so on.}

\item{benchmark.gains}{Numeric vector or matrix of gains for 1 or 2 benchmark 
indexes. Only used if \code{x.metric} and/or \code{y.metric} require 
benchmark indexes to calculate. For example, to plot correlation with SPY on 
the x-axis and correlation with TLT on the y-axis, set 
\code{x.metric = "pearson"} and \code{y.metric = "pearson2"}, and 
input \code{benchmark.gains} as a 2-column matrix of gains for SPY and TLT.}

\item{reference.gains}{Numeric vector or matrix of gains for funds to 
include on graph as data points for comparative purposes.}

\item{step.data}{Numeric value specifying allocation increments for data used 
to fit curves.}

\item{step.points}{Numeric value specifying allocation increments for data 
points on top of curves. Set to \code{NULL} to suppress data points.}

\item{step.curves}{Numeric value specifying allocation increments for first 
fund in each set. For example, set to \code{0.2} for six curves for each 
three-fund set: one for 0\% fund 1, another for 20\% fund 1, and so on, up to 
100\% fund 1 (which is technically a data point, not a curve).}

\item{x.metric}{Character string specifying performance metric to plot on the 
x-axis. Choices are: 

\code{"mean"} or \code{"sd"} for mean or standard deviation of gains 

\code{"growth"} or \code{"cagr"} for total or annualized growth 

\code{"mdd"} for maximum drawdown

\code{"sharpe"} or \code{"sortino"} for Sharpe or Sortino ratio 

\code{"alpha"}, \code{"beta"}, or \code{"r.squared"} for those metrics from a 
fitted linear regression on benchmark fund

\code{"pearson"} or \code{"spearman"} for Pearson or Spearman correlation 
with benchmark fund

\code{"alpha2"}, \code{"beta2"}, \code{"r.squared2"}, \code{"pearson2"}, or 
\code{"spearman2"} for same as previously described, but using the second 
benchmark index

\code{"auto.pearson"} or \code{"auto.spearman"} for Pearson or Spearman 
autocorrelation, defined as the correlation between subsequent gains
  
\code{"allocation"} for allocation to first fund in each pair.}

\item{y.metric}{Same as \code{x.metric}, but for y-axis variable.}

\item{tickerlabel.offsets}{Either a numeric vector of length 2 giving the 
x- and y-axis offsets for ticker labels, or a 2-column matrix where each row 
gives the x- and y-axis offsets for a particular ticker, following the same 
order of tickers as in the \code{tickers} matrix (going down each column, 
then to the next row) or across the \code{ticker.gains} matrix.}

\item{reflabel.offsets}{Either a numeric vector of length 2 giving the x- and 
y-axis offsets for reference ticker labels, or a 2-column matrix where each 
row gives the x- and y-axis offsets for a particular reference ticker, 
following the same order of tickers as in the \code{reference.tickers} vector 
or across the \code{reference.gains} matrix.}

\item{add.plot}{Logical value for whether to add plot data to the current 
plot frame, as opposed to opening a new one.}

\item{colors}{Character string of colors for each fund pair.}

\item{plot.list}{List of arguments to pass to 
\code{\link[graphics]{plot}}, \code{\link[graphics]{points}}, and 
\code{\link[graphics]{text}}.}

\item{points.list}{List of arguments to pass to 
\code{\link[graphics]{plot}}, \code{\link[graphics]{points}}, and 
\code{\link[graphics]{text}}.}

\item{text.list}{List of arguments to pass to 
\code{\link[graphics]{plot}}, \code{\link[graphics]{points}}, and 
\code{\link[graphics]{text}}.}

\item{pdf.list}{List of arguments to 
pass to \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{bmp}}, etc. For 
each that is not \code{NULL}, the corresponding figure file gets created.}

\item{bmp.list}{List of arguments to 
pass to \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{bmp}}, etc. For 
each that is not \code{NULL}, the corresponding figure file gets created.}

\item{jpeg.list}{List of arguments to 
pass to \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{bmp}}, etc. For 
each that is not \code{NULL}, the corresponding figure file gets created.}

\item{png.list}{List of arguments to 
pass to \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{bmp}}, etc. For 
each that is not \code{NULL}, the corresponding figure file gets created.}

\item{tiff.list}{List of arguments to 
pass to \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{bmp}}, etc. For 
each that is not \code{NULL}, the corresponding figure file gets created.}
}
\value{
In addition to the graph, a list containing the following: 
\enumerate{
\item List named \code{portfolio.xy} where each element is a list of 
two-column matrices of x- and y-axis values for a particular three-fund set 
with a particular allocation to fund 1.
\item Numeric vector named \code{means} with mean gains for each fund.
\item Numeric matrix named \code{corr.matrix} with a correlation matrix for 
gains for each fund.
}
}
\description{
Useful for visualizing performance of three-fund portfolios, typically by 
plotting a measure of growth vs. a measure of volatility.
}
\examples{
# Plot mean vs. SD for UPRO/VBLTX/VWEHX portfolio, and compare to VFINX and 
# BRK-B
fig1 <- threefunds_graph(tickers = c("VWEHX", "VBLTX", "UPRO"), 
                         reference.tickers = c("VFINX", "BRK-B"))
                      
# Same funds, but annualized growth vs. maximum drawdown
fig2 <- threefunds_graph(tickers = c("VWEHX", "VBLTX", "UPRO"), 
                         reference.tickers = c("VFINX", "BRK-B"),
                         x.metric = "mdd", y.metric = "cagr")

}
\references{
Ryan, J.A. and Ulrich, J.M. (2017) quantmod: Quantitative Financial Modelling 
Framework. R package version 0.4-12, 
\url{https://CRAN.R-project.org/package=quantmod}.
}
