% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_metric.R
\name{title_metric}
\alias{title_metric}
\title{Convert Title back to Performance Metric}
\usage{
title_metric(title)
}
\arguments{
\item{title}{Character string.}
}
\value{
Character string.
}
\description{
For internal use only.
}
