% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_etfs}
\alias{vanguard_etfs}
\alias{vanguard.etfs}
\title{Vanguard ETF's}
\source{
\url{https://investor.vanguard.com/mutual-funds/list#/mutual-funds/asset-class/month-end-returns}
}
\description{
Vanguard ETF's
}
