% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_funds}
\alias{vanguard_funds}
\alias{vanguard.funds}
\title{Vanguard Mutual Funds}
\source{
\url{https://investor.vanguard.com/mutual-funds/list#/mutual-funds/asset-class/month-end-returns}
}
\description{
Vanguard Mutual Funds
}
