% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_products}
\alias{vanguard_products}
\alias{vanguard.products}
\title{Vanguard Products}
\source{
\url{https://investor.vanguard.com/mutual-funds/list#/mutual-funds/asset-class/month-end-returns}


Lookup Table for Performance Metrics

Original
}
\description{
Vanguard Products
}
