\name{vanguard.sector.etfs}
\alias{vanguard.sector.etfs}
\docType{data}
\title{
List of Vanguard Sector & Specialty ETFs as of April 2, 2017
}
\description{
Simply a character vector of ticker symbols for Vanguard ETFs.
}
\format{
  A character vector with 11 tickers.
}
\source{
\url{https://investor.vanguard.com/etf/list#/etf/asset-class/month-end-returns}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}