\name{vanguard.stock.etfs}
\alias{vanguard.stock.etfs}
\docType{data}
\title{
List of Vanguard Stock ETFs as of March 6, 2017.
}
\description{
Simply a character vector of ticker symbols for Vanguard ETFs.
}
\usage{data("vanguard.stock.etfs")}
\format{
  A character vector with 17 tickers.
}
\source{
\url{https://investor.vanguard.com/etf/list?assetclass=stk#/etf/asset-class/month-end-returns}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
# Load and print list of ETFs
#data(vanguard.stock.etfs)
#vanguard.stock.etfs
}
\keyword{datasets}