\name{vanguard.targetdate.funds}
\alias{vanguard.targetdate.funds}
\docType{data}
\title{
List of Vanguard Target Date Mutual Funds as of March 6, 2017.
}
\description{
Simply a character vector of ticker symbols for Vanguard mutual funds.
}
%\usage{data("vanguard.targetdate.funds")}
\format{
  A character vector with 12 tickers.
}
\source{
\url{https://investor.vanguard.com/mutual-funds/list#/mutual-funds/asset-class/month-end-returns}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}