% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_taxexempt_bond_funds}
\alias{vanguard_taxexempt_bond_funds}
\title{Vanguard Tax-exempt Bond Mutual Funds and Inception Dates}
\source{
\url{https://investor.vanguard.com/etf/list?assetclass=bond#/etf/asset-class/month-end-returns}
}
\description{
Vanguard Tax-exempt Bond Mutual Funds and Inception Dates
}
