\name{zzz}
\alias{zzz}
\docType{data}
\title{
Daily Closing Prices for Imaginary Stock ZZZ Over 10 Years
}
\description{
This dataset is included to help illustrate various functions in the stocks package.
}
\usage{data("zzz")}
\format{
  The format is:
 num [1:2511] 37.5 36.2 35.9 36.4 36.6 ...
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\keyword{datasets}