% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_trailing50.R
\name{beta_trailing50}
\alias{beta_trailing50}
\title{Beta for Last 50 Daily Gains}
\usage{
beta_trailing50(ticker, bench = "SPY", ...)
}
\arguments{
\item{ticker}{Character string with ticker symbols that Yahoo! Finance 
recognizes.}

\item{bench}{Character string with ticker symbol for benchmark.}

\item{...}{Arguments to pass to \code{\link{load_gains}}.}
}
\value{
Numeric value.
}
\description{
Calculates beta for a ticker symbol based on the previous 50 daily gains.
}
\examples{
\dontrun{
# Calculate TLT's beta based on the previous 50 daily gains
beta_trailing50("TLT")
}


}
\references{
Ryan, J.A. and Ulrich, J.M. (2017) quantmod: Quantitative Financial Modelling 
Framework. R package version 0.4-12, 
\url{https://CRAN.R-project.org/package=quantmod}.
}
