% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metric.R
\name{calc_metric}
\alias{calc_metric}
\title{Calculate Performance Metric}
\usage{
calc_metric(gains, metric = "mean", units.year = 252,
  benchmark.gains = NULL)
}
\arguments{
\item{gains}{Numeric vector.}

\item{metric}{Character string specifying metric to calculate. Choices are
\code{"mean"}, \code{"sd"}, \code{"growth.x"} for growth of $x where x is the
initial value, \code{"growth"} for percent growth, \code{"cagr"} for compound
annualized growth rate, \code{"mdd"} for max drawdown, \code{"sharpe"},
\code{"sortino"}, \code{"alpha"}, \code{"alpha.annualized"}, \code{"beta"},
\code{"r.squared"}, \code{"pearson"} or \code{"spearman"} for
Pearson/Spearman correlation with benchmark, and \code{"auto.pearson"} or
\code{"auto.spearman"} for Pearson/Spearman autocorrelation.}

\item{units.year}{Integer value.}

\item{benchmark.gains}{Numeric vector.}
}
\value{
Numeric value.
}
\description{
Mainly a helper function for \code{\link{plot_metrics_overtime}}.
}
