% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculate Performance Metrics}
\usage{
calc_metrics(gains = NULL, metrics = c("mean", "sd", "growth.10k",
  "growth", "cagr", "mdd", "sharpe", "sortino", "alpha",
  "alpha.annualized", "beta", "r.squared", "pearson", "spearman",
  "auto.pearson", "auto.spearman"), prices = NULL, tickers = NULL, ...,
  benchmark = "SPY")
}
\arguments{
\item{gains}{Data frame with one column of gains for each investment and a
date variable named Date.}

\item{metrics}{Character vector specifying metrics to calculate. Choices are
\code{"mean"}, \code{"sd"}, \code{"growth.x"} for growth of $x where x is the
initial value, \code{"growth"} for percent growth, \code{"cagr"} for compound
annualized growth rate, \code{"mdd"} for max drawdown, \code{"sharpe"},
\code{"sortino"}, \code{"alpha"}, \code{"alpha.annualized"}, \code{"beta"},
\code{"r.squared"}, \code{"pearson"} or \code{"spearman"} for
Pearson/Spearman correlation with benchmark, and \code{"auto.pearson"} or
\code{"auto.spearman"} for Pearson/Spearman autocorrelation.}

\item{prices}{Data frame with one column of prices for each investment and a
date variable named Date.}

\item{tickers}{Character vector of ticker symbols that Yahoo! Finance
recognizes, if you want to download data on the fly.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{benchmark}{Character string specifying which fund to use as a
benchmark for metrics that require one.}
}
\value{
Data frame with performance metrics for each investment.
}
\description{
Useful for comparing funds on one or more metrics.
}
\examples{
\dontrun{
# Calculate performance metrics for FANG stocks since the beginning of 2019
calc_metrics(tickers = fang, from = "2019-01-01")

# Repeat, but use step-by-step approach with piping (need SPY to calculate
# alpha and beta)
c("SPY", fang) \%>\%
  load_gains(from = "2019-01-01") \%>\%
  calc_metrics()
}


}
