% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics_123.R
\name{calc_metrics_123}
\alias{calc_metrics_123}
\title{Calculate Performance Metrics for Any Combination of Individual Funds,
2-Fund Portfolios, and 3-Fund Portfolios}
\usage{
calc_metrics_123(
  gains = NULL,
  metrics = c("mean", "sd"),
  tickers = NULL,
  ...,
  step = 1,
  prices = NULL,
  benchmark = "SPY"
)
}
\arguments{
\item{gains}{Data frame with a date variable named Date and one column of
gains for each fund.}

\item{metrics}{Character vector specifying metrics to calculate. See
\code{?calc_metrics} for choices.}

\item{tickers}{List where each element is a character vector of ticker
symbols for a particular fund combination, e.g.
\code{list("BRK-B", c("SPY", "TLT")}. Each set can contain 1-3 funds.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{step}{Numeric value specifying fund allocation increments.}

\item{prices}{Data frame with a date variable named Date and one column of
prices for each fund.}

\item{benchmark}{Character string specifying which fund to use as a
benchmark for metrics that require one.}
}
\value{
Data frame with performance metrics for each portfolio at each allocation.
}
\description{
Integrates \code{calc_metrics}, \code{calc_metrics_2funds}, and
\code{calc_metrics_3funds} into a single function, so you can compare
strategies of varying complexities.
}
\examples{
\dontrun{
# Calculate CAGR vs. max drawdown for BRK-B, SPY/TLT, and VWEHX/VBLTX/VFINX
df <- calc_metrics_123(
  tickers = list("BRK-B", c("SPY", "TLT"), c("VWEHX", "VBLTX", "VFINX")),
  metrics = c("cagr", "mdd")
)
head(df)

# To plot, just pipe into plot_metrics_123
df \%>\%
  plot_metrics_123()

# Or bypass calc_metrics_123 altogether
plot_metrics_123(
  formula = cagr ~ mdd,
  tickers = list("BRK-B", c("SPY", "TLT"), c("VWEHX", "VBLTX", "VFINX"))
)
}


}
