% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics_2funds.R
\name{calc_metrics_2funds}
\alias{calc_metrics_2funds}
\title{Calculate Performance Metrics for 2-Fund Portfolios with Varying
Allocations}
\usage{
calc_metrics_2funds(gains = NULL, metrics = c("mean", "sd"),
  tickers = NULL, ..., prices = NULL, benchmark = "SPY",
  ref.tickers = NULL)
}
\arguments{
\item{gains}{Data frame with a date variable named Date and one column of
gains for each fund.}

\item{metrics}{Character vector specifying metrics to calculate. See
\code{?calc_metrics} for choices.}

\item{tickers}{Character vector of ticker symbols, where the first two are
are a 2-fund pair, the next two are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{prices}{Data frame with a date variable named Date and one column of
prices for each fund.}

\item{benchmark}{Character string specifying which fund to use as a
benchmark for metrics that require one.}

\item{ref.tickers}{Character vector of ticker symbols to include.}
}
\value{
Data frame with performance metrics for each portfolio at each allocation.
}
\description{
Useful for assessing the characteristics of 2-fund portfolios.
}
\examples{
\dontrun{
# Calculate CAGR and max drawdown for UPRO/VBLTX
df <- calc_metrics_2funds(
  metrics = c("cagr", "mdd"),
  tickers = c("UPRO", "VBLTX")
)
head(df)

# To plot, just pipe into plot_metrics_2funds
df \%>\%
  plot_metrics_2funds()

# Or bypass calc_metrics_2funds altogether
plot_metrics_2funds(
  formula = cagr ~ mdd,
  tickers = c("UPRO", "VBLTX")
)
}


}
