% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics_3funds.R
\name{calc_metrics_3funds}
\alias{calc_metrics_3funds}
\title{Calculate Performance Metrics for Three-Fund Portfolios with Varying 
Allocations}
\usage{
calc_metrics_3funds(gains = NULL, metrics = c("mean", "sd"),
  tickers = NULL, ..., step1 = 5, step2 = step1, prices = NULL,
  benchmark = "SPY", ref.tickers = "SPY")
}
\arguments{
\item{gains}{Data frame with a date variable named Date and one column of 
gains for each investment.}

\item{metrics}{Character vector specifying metrics to calculate. See 
\code{?calc_metrics} for choices.}

\item{tickers}{Character vector of ticker symbols, where the first three are 
are a three-fund set, the next three are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to 
\code{\link{load_gains}}.}

\item{step1}{Numeric value controlling allocation increments for first fund 
of each set.}

\item{step2}{Numeric value controlling allocation increments for the second 
and third fund in each set.}

\item{prices}{Data frame with a date variable named Date and one column of 
prices for each investment.}

\item{benchmark}{Character string specifying which fund to use as a 
benchmark for metrics that require one.}

\item{ref.tickers}{Character vector of ticker symbols to include.}
}
\value{
Depending on \code{return}, a \code{\link[ggplot2]{ggplot}} object, a data
frame, or a list containing both.
}
\description{
Useful for assessing the characteristics of three-fund portfolios.
}
\examples{
\dontrun{
# Plot mean vs. SD for UPRO/VBLTX/VWEHX
plot_metrics_3funds(mean ~ sd, tickers = c("UPRO", "VBLTX", "VWEHX"))

# Plot CAGR vs. MDD for FB/AAPL/NFLX and SPY/TLT/JNK
plot_metrics_3funds(cagr ~ mdd, tickers = c("FB", "AAPL", "NFLX", "SPY", "TLT", "JNK"))

# Plot Sharpe ratio vs. allocation for the same sets
plot_metrics_3funds(sharpe ~ allocation, tickers = c("FB", "AAPL", "NFLX", "SPY", "TLT", "JNK"))
}


}
