% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Financial_Ratios.R
\name{filter_company_data}
\alias{filter_company_data}
\title{Quick Look at a Company's Financials}
\usage{
filter_company_data(tik)
}
\arguments{
\item{tik}{A character string consisting of the stock ticker for the company.}
}
\value{
A dataframe containing financial ratios for the company if it is a public stock on NASDAQ or NYSE.
}
\description{
This function allows a user to view several financial ratios for a company.
}
\examples{
#Example of `filter_company_data` in use is in the stock package in action section
}
