% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gains_prices.R
\name{gains_prices}
\alias{gains_prices}
\title{Convert Sequence of Gains to Sequence of Prices}
\usage{
gains_prices(gains, initial = 10000, date1 = NULL)
}
\arguments{
\item{gains}{Numeric vector of gains for one investment, or data frame with 
one column for each investment and an optional Date variable.}

\item{initial}{Numeric value.}

\item{date1}{Date to use for initial price.}
}
\value{
Numeric vector or data frame.
}
\description{
Converts sequence of gains and initial balance to sequence of prices for one 
or more investments.
}
\examples{
# Simulate daily gains over a 5-year period
set.seed(123)
gains <- rnorm(n = 252 * 5, mean = 0.001, sd = 0.02)

# Plot balance over time if initial balance is $10,000
prices <- gains_prices(gains)
plot(prices)


}
