\name{highyield.etfs}
\alias{highyield.etfs}
\docType{data}
\title{
List of High Yield ETFs from ETFdb.com as of April 7, 2017
}
\description{
Simply a character vector of ticker symbols for high-yield ETFs.
}
\format{
  A character vector with 46 tickers.
}
\source{
\url{http://etfdb.com/etfdb-category/high-yield-bonds/#etfs&sort_name=assets_under_management&sort_order=desc&page=2}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}