\name{highyield.etfs}
\alias{highyield.etfs}
\docType{data}
\title{
High-Yield ETFs from ETFdb.com and Inception Dates as of April 7, 2017
}
\description{
Data frame where first column is ticker symbol and second column is inception
date.
}
\format{
  A 46-row data frame with 2 variables: \code{ticker} (character string) and
  \code{start.date} (date).
}
\source{
\url{http://etfdb.com/etfdb-category/high-yield-bonds/#etfs&sort_name=assets_under_management&sort_order=desc&page=2}
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}
