% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{highyield_etfs}
\alias{highyield_etfs}
\title{High-Yield ETFs from ETFdb.com and Inception Dates}
\source{
\url{http://etfdb.com/etfdb-category/high-yield-bonds/#etfs&sort_name=assets_under_management&sort_order=desc&page=2}
}
\description{
High-Yield ETFs from ETFdb.com and Inception Dates
}
