% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{largest_etfs}
\alias{largest_etfs}
\title{Largest 100 Market Cap ETFs (as of 3/2/18) and Inception Dates}
\source{
\url{http://etfdb.com/compare/market-cap/}
}
\description{
Largest 100 Market Cap ETFs (as of 3/2/18) and Inception Dates
}
