% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot Investment Growth}
\usage{
plot_growth(prices = NULL, tickers = NULL, ..., gains = NULL,
  initial = 10000, plotly = FALSE, title = "Growth Over Time",
  base_size = 16, tooltip_size = 20, point_size = 1, line_size = 1,
  ticklabel_size = 8, return = "plot")
}
\arguments{
\item{prices}{Data frame with one column of prices for each investment and a
date variable named Date.}

\item{tickers}{Character vector of ticker symbols that Yahoo! Finance
recognizes, if you want to download data on the fly.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{gains}{Data frame with one column of gains for each investment and a
date variable named Date.}

\item{initial}{Numeric value specifying value to scale initial prices to.}

\item{plotly}{Logical value for whether to convert the
\code{\link[ggplot2]{ggplot}} to a \code{\link[plotly]{plotly}} object
internally.}

\item{title}{Character string.}

\item{base_size}{Numeric value to pass to
\code{\link[ggplot2:ggtheme]{theme_gray}}.}

\item{tooltip_size}{Numeric value to pass to \code{\link[plotly]{style}}.}

\item{point_size}{Numeric value to pass to \code{\link[ggplot2]{geom_point}}.}

\item{line_size}{Numeric value to pass to
\code{\link[ggplot2:geom_path]{geom_line}}.}

\item{ticklabel_size}{Numeric value to pass to \code{\link[ggplot2]{theme}}.}

\item{return}{Character string specifying what to return. Choices are
\code{"plot"}, \code{"data"}, and \code{"both"}.}
}
\value{
Depending on \code{return} and \code{plotly}, a
\code{\link[ggplot2]{ggplot}}/\code{\link[plotly]{plotly}} object, a data
frame with the source data, or a list containing both.

A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Useful for comparing the performance of several investments, over their full
histories or mutual lifetimes.
}
\examples{
\dontrun{
# Plot growth of $10k in VFINX and BRK-B
plot_growth(tickers = c("VFINX", "BRK-B"))
}

}
