% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics_123.R
\name{plot_metrics_123}
\alias{plot_metrics_123}
\title{Plot One Performance Metric vs. Another for Any Number of Single Funds,
2-Fund Portfolios, and 3-Fund Portfolios}
\usage{
plot_metrics_123(metrics = NULL, formula = mean ~ sd, tickers = NULL,
  ..., step = 1, gains = NULL, prices = NULL, benchmark = "SPY",
  y.benchmark = benchmark, x.benchmark = benchmark, plotly = FALSE,
  title = NULL, base_size = 16, label_size = 5, return = "plot")
}
\arguments{
\item{metrics}{Data frame with Fund column and column for each metric you
want to plot. Typically the result of a prior call to
\code{\link{calc_metrics_123}}.}

\item{formula}{Formula specifying what to plot, e.g. \code{mean ~ sd},
\code{cagr ~ mdd}, or \code{sharpe ~ allocation}. See \code{?calc_metrics}
for list of metrics to choose from (\code{"allocation"} is an extra option
here). If you specify \code{metrics}, default behavior is to use
\code{mean ~ sd} unless either is not available, in which case the first two
performance metrics that appear as columns in \code{metrics} are used.}

\item{tickers}{Character vector of ticker symbols, where the first three are
are a three-fund set, the next three are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{step}{Numeric value specifying fund allocation increments.}

\item{gains}{Data frame with a date variable named Date and one column of
gains for each fund.}

\item{prices}{Data frame with a date variable named Date and one column of
prices for each fund.}

\item{benchmark, y.benchmark, x.benchmark}{Character string specifying which
fund to use as benchmark for metrics (if you request \code{alpha},
\code{alpha.annualized}, \code{beta}, or \code{r.squared}).}

\item{plotly}{Logical value for whether to convert the
\code{\link[ggplot2]{ggplot}} to a \code{\link[plotly]{plotly}} object
internally.}

\item{title}{Character string.}

\item{base_size}{Numeric value.}

\item{label_size}{Numeric value.}

\item{return}{Character string specifying what to return. Choices are
\code{"plot"}, \code{"data"}, and \code{"both"}.}
}
\value{
Depending on \code{return}, a \code{\link[ggplot2]{ggplot}} object, a data
frame, or a list containing both.
}
\description{
Integrates \code{plot_metrics}, \code{plot_metrics_2funds}, and
\code{plot_metrics_3funds} into a single function, so you can visualize
strategies of varying complexities on one figure.
}
\details{
If you prefer to have complete control over the plotting, you can set
\code{return = "data"} to just get the source data.
}
\examples{
\dontrun{
# Plot CAGR vs. max drawdown for BRK-B, SPY/TLT, and VWEHX/VBLTX/VFINX
plot_metrics_123(
  formula = cagr ~ mdd,
  tickers = list("BRK-B", c("SPY", "TLT"), c("VWEHX", "VBLTX", "VFINX"))
)
}



}
