% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics_2funds.R
\name{plot_metrics_2funds}
\alias{plot_metrics_2funds}
\title{Plot One Performance Metric vs. Another for 2-Fund Portfolios}
\usage{
plot_metrics_2funds(
  metrics = NULL,
  formula = mean ~ sd,
  tickers = NULL,
  ...,
  points = seq(0, 100, 10),
  gains = NULL,
  prices = NULL,
  benchmark = "SPY",
  y.benchmark = benchmark,
  x.benchmark = benchmark,
  ref.tickers = NULL,
  plotly = FALSE,
  title = NULL,
  base_size = 16,
  label_size = 5,
  return = "plot"
)
}
\arguments{
\item{metrics}{Data frame with Fund column and column for each metric you
want to plot. Typically the result of a prior call to
\code{\link{calc_metrics_2funds}}.}

\item{formula}{Formula specifying what to plot, e.g. \code{mean ~ sd},
\code{cagr ~ mdd}, or \code{sharpe ~ allocation}. See \code{?calc_metrics}
for list of metrics to choose from (\code{"allocation"} is an extra option
here). If you specify \code{metrics}, default behavior is to use
\code{mean ~ sd} unless either is not available, in which case the first two
performance metrics that appear as columns in \code{metrics} are used.}

\item{tickers}{Character vector of ticker symbols, where the first two are
are a two-fund pair, the next two are another, and so on.}

\item{...}{Arguments to pass along with \code{tickers} to
\code{\link{load_gains}}.}

\item{points}{Numeric vector specifying allocations to include as points on
the curve. Set to \code{NULL} for none (0 and 100 will still be included).}

\item{gains}{Data frame with a date variable named Date and one column of
gains for each fund.}

\item{prices}{Data frame with a date variable named Date and one column of
prices for each fund.}

\item{benchmark, y.benchmark, x.benchmark}{Character string specifying which
fund to use as benchmark for metrics (if you request \code{alpha},
\code{alpha.annualized}, \code{beta}, or \code{r.squared}).}

\item{ref.tickers}{Character vector of ticker symbols to include on the
plot.}

\item{plotly}{Logical value for whether to convert the
\code{\link[ggplot2]{ggplot}} to a \code{\link[plotly]{plotly}} object
internally.}

\item{title}{Character string.}

\item{base_size}{Numeric value.}

\item{label_size}{Numeric value.}

\item{return}{Character string specifying what to return. Choices are
\code{"plot"}, \code{"data"}, and \code{"both"}.}
}
\value{
Depending on \code{return}, a \code{\link[ggplot2]{ggplot}} object, a data
frame, or a list containing both.
}
\description{
Useful for visualizing the behavior of 2-fund portfolios, e.g. by plotting
a measure of growth vs. a measure of volatility.
}
\examples{
\dontrun{
# Plot mean vs. SD for UPRO/VBLTX, and compare to SPY
plot_metrics_2funds(
  formula = mean ~ sd,
  tickers = c("UPRO", "VBLTX")
)

# Plot CAGR vs. max drawdown for AAPL/GOOG and FB/TWTR
plot_metrics_2funds(
  formula = cagr ~ mdd,
  tickers = c("AAPL", "GOOG", "FB", "TWTR")
)

# Plot Sharpe ratio vs. allocation for SPY/TLT
plot_metrics_2funds(
  formula = sharpe ~ allocation,
  tickers = c("SPY", "TLT")
)
}

}
