\name{sector.spdr.etfs}
\alias{sector.spdr.etfs}
\docType{data}
\title{
List of Sector SPDR ETFs as of April 7, 2017
}
\description{
Simply a character vector of ticker symbols for sector SPDR ETFs.
}
\format{
  A character vector with 10 tickers.
}
\source{
\url{http://www.sectorspdr.com/sectorspdr/sectors/performance}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}