% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{sector_spdr_etfs}
\alias{sector_spdr_etfs}
\title{Sector SPDR ETFs and Inception Dates}
\source{
\url{http://www.sectorspdr.com/sectorspdr/sectors/performance}
}
\description{
Sector SPDR ETFs and Inception Dates
}
