% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{package}
\name{stocks}
\alias{stocks}
\alias{stocks-package}
\title{Stock Market Analysis}
\description{
Functions for analyzing and visualizing stock market data. Main features are
loading and aligning historical data, calculating performance metrics for
individual funds or portfolios (e.g. annualized growth, maximum drawdown,
Sharpe/Sortino ratio), and creating graphs.
}
\details{
\tabular{ll}{
Package: \tab stocks \cr
Type: \tab Package \cr
Version: \tab 2.0.0 \cr
Date: \tab 2020-04-09 \cr
License: \tab GPL-3 \cr
}

See \href{https://cran.r-project.org/package=stocks}{CRAN documentation} for
full list of functions and the
\href{https://github.com/vandomed/stocks}{GitHub page} for an overview of the
package with some examples.
}
\references{
Jeffrey A. Ryan and Joshua M. Ulrich (2019). quantmod: Quantitative Financial
Modelling Framework. R package version 0.4-15.
\url{https://CRAN.R-project.org/package=quantmod}
}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
