% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_bond_funds}
\alias{vanguard_bond_funds}
\title{Vanguard Bond Mutual Funds}
\source{
\url{https://investor.vanguard.com/mutual-funds/list?assetclass=bond#/mutual-funds/asset-class/month-end-returns}
}
\description{
Vanguard Bond Mutual Funds
}
