\name{vanguard.etfs}
\alias{vanguard.etfs}
\docType{data}
\title{
List of Vanguard ETFs and their inception dates, current as of March 20, 2015.
}
\description{
Just a character vector of ticker symbols for Vanguard ETFs.
}
\usage{data("vanguard.etfs")}
\format{
  A data frame with 67 observations on the following 2 variables.
  \describe{
    \item{\code{ticker}}{a factor with levels \code{BIV} \code{BLV} \code{BND} \code{BNDX} \code{BSV} \code{EDV} \code{IVOG} \code{IVOO} \code{IVOV} \code{MGC} \code{MGK} \code{MGV} \code{VAW} \code{VB} \code{VBK} \code{VBR} \code{VCIT} \code{VCLT} \code{VCR} \code{VCSH} \code{VDC} \code{VDE} \code{VEA} \code{VEU} \code{VFH} \code{VGIT} \code{VGK} \code{VGLT} \code{VGSH} \code{VGT} \code{VHT} \code{VIG} \code{VIOG} \code{VIOO} \code{VIOV} \code{VIS} \code{VMBS} \code{VNQ} \code{VNQI} \code{VO} \code{VOE} \code{VONE} \code{VONG} \code{VONV} \code{VOO} \code{VOOG} \code{VOOV} \code{VOT} \code{VOX} \code{VPL} \code{VPU} \code{VSS} \code{VT} \code{VTHR} \code{VTI} \code{VTIP} \code{VTV} \code{VTWG} \code{VTWO} \code{VTWV} \code{VUG} \code{VV} \code{VWO} \code{VWOB} \code{VXF} \code{VXUS} \code{VYM}}
    \item{\code{start.date}}{a factor with levels \code{2001-06-15} \code{2002-01-04} \code{2004-01-30} \code{2004-09-29} \code{2005-03-10} \code{2006-05-02} \code{2006-08-25} \code{2006-11-16} \code{2007-03-08} \code{2007-04-10} \code{2007-07-26} \code{2007-12-27} \code{2008-01-29} \code{2008-06-26} \code{2009-04-06} \code{2009-11-23} \code{2010-01-04} \code{2010-09-09} \code{2010-09-22} \code{2010-11-01} \code{2011-01-28} \code{2012-10-16} \code{2013-06-04}}
  }
}
\source{
\url{https://personal.vanguard.com/us/funds/etf/ETFTableView}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}