% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_international_funds}
\alias{vanguard_international_funds}
\title{Vanguard International Mutual Funds and Inception Dates}
\source{
\url{https://investor.vanguard.com/mutual-funds/list?assetclass=bond#/mutual-funds/asset-class/month-end-returns}
}
\description{
Vanguard International Mutual Funds and Inception Dates
}
