\name{vanguard.midcap.funds}
\alias{vanguard.midcap.funds}
\docType{data}
\title{
List of Vanguard Mid-cap Stock Mutual Funds as of April 2, 2017
}
\description{
Simply a character vector of ticker symbols for Vanguard mutual funds.
}
\format{
  A character vector with 9 tickers.
}
\source{
\url{https://investor.vanguard.com/mutual-funds/list#/mutual-funds/asset-class/month-end-returns}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}