% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_sector_etfs}
\alias{vanguard_sector_etfs}
\title{Vanguard Sector & Specialty ETFs and Inception Dates}
\source{
\url{https://investor.vanguard.com/etf/list?assetclass=bond#/etf/asset-class/month-end-returns}
}
\description{
Vanguard Sector & Specialty ETFs and Inception Dates
}
