\name{vanguard.smallcap.etfs}
\alias{vanguard.smallcap.etfs}
\docType{data}
\title{
Vanguard Small-cap Stock ETFs and Inception Dates as of April 2, 2017
}
\description{
Data frame where first column is ticker symbol and second column is inception date.
}
\format{
  A 3-row data frame with 2 variables: ticker (character string) and start.date (date).
}
\source{
\url{https://investor.vanguard.com/etf/list?assetclass=stk#/etf/asset-class/month-end-returns}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}