\name{vanguard.targetdate.funds}
\alias{vanguard.targetdate.funds}
\docType{data}
\title{
Vanguard Target Date Mutual Funds as of April 2, 2017
}
\description{
Data frame where first column is ticker symbol and second column is inception date.
}
\format{
  A 12-row data frame with 2 variables: \code{ticker} (character string) and \code{start.date} (date).
}
\source{
\url{https://investor.vanguard.com/mutual-funds/list#/mutual-funds/asset-class/month-end-returns}
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}