% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_targetdate_funds}
\alias{vanguard_targetdate_funds}
\title{Vanguard Target Date Mutual Funds}
\source{
\url{https://investor.vanguard.com/etf/list?assetclass=bond#/etf/asset-class/month-end-returns}
}
\description{
Vanguard Target Date Mutual Funds
}
