% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks-data.R
\docType{data}
\name{vanguard_traditional_funds}
\alias{vanguard_traditional_funds}
\title{Vanguard Traditional Mutual Funds and Inception Dates}
\source{
\url{https://investor.vanguard.com/etf/list?assetclass=bond#/etf/asset-class/month-end-returns}
}
\description{
Vanguard Traditional Mutual Funds and Inception Dates
}
